Add more tests for api route
